# Faster boot time for commercials