--- a/static/src/javascripts/projects/common/modules/experiments/tests/live-blog-chrome-notifications-prod.js
+++ b/static/src/javascripts/projects/common/modules/experiments/tests/live-blog-chrome-notifications-prod.js
@@ -1,9 +1,7 @@
 define([
-    'common/utils/$',
     'common/utils/config',
     'common/utils/detect'
 ], function (
-    $,
     config,
     detect
 ) {