--- a/static/src/javascripts/projects/common/modules/ui/images.js
+++ b/static/src/javascripts/projects/common/modules/ui/images.js
@@ -1,23 +1,18 @@
 define([
-    'bonzo',
+    'qwery',
     'picturefill',
-    'common/utils/$',
-    'common/utils/$css',
     'common/utils/mediator'
 ],
 function (
-    bonzo,
+    qwery,
     picturefill,
-    $,
-    $css,
     mediator
 ) {
 
     var images = {
 
         upgradePictures: function (context) {
-            var images = [].slice.call($('img[srcset], picture img', context || document.body));
-            picturefill({ elements: images });
+            picturefill({ elements: qwery('img[srcset], picture img', context || document.body) });
         },
 
         listen: function () {