… common/app/views/fragments/metaData.scala.html
… static/src/javascripts/projects/common/modules/experiments/tests/loyal-adblocking-survey.js
… static/src/javascripts/projects/common/utils/cookies.js
… static/src/javascripts/projects/common/utils/detect.js
… static/src/javascripts/projects/common/utils/get-property.js
… static/src/javascripts/projects/common/utils/robust.js
… static/src/javascripts/projects/common/utils/storage.js
… static/src/javascripts/projects/common/utils/url.js
… static/src/javascripts/projects/common/views/commercial/survey/survey-simple.html
… static/src/javascripts/projects/facia/modules/ui/container-show-more.js