cc @jbreckmckye @uplne @sndrs @gtrufitt 

We have a big problem. On an empty cache, commercials load very, very late. The GPT library necessary to make any ad request is only loaded around the 4s mark:

![picture 3](https://cloud.githubusercontent.com/assets/629976/15424161/05b79970-1e79-11e6-8a67-eca756530539.jpg)

Still after the load event on a primed cache:

![picture 4](https://cloud.githubusercontent.com/assets/629976/15424205/448f84aa-1e79-11e6-951c-d37286bf06f1.jpg)

This delay is noticeable when you browse the site and focus on the ads (admittedly, nobody but us do). The problem is that this delay has a very tangible effect our the bottom line—which we are finally able to measure thanks to @JonNorman's and @kenlim's project Maguire.

Using performance timing on my local machine, I found that:

* js kicks in around the **1.5s** mark
* standard js starts **1.5s** after that and takes only 5–10ms to run
* commercial js starts **1.5s** after that and runs for **~400ms** 😭 

Even if we take away the cost of doing network requests in dev mode, it's still pretty obvious we have some contention points here:

* the time necessary to parse the DOM and run blocking JS is huge
* the time necessary to load and parse all the standard and commercial JS dependencies are huge
* commercial is f-ing slow

We have to do something about it, and this PR is the first attempt in a series at doing so. Essentially I think along those lines:

* html/js is optimized for current browsers first
* older browsers will incur additional http requests to get the polyfills/helpers they need, depending on which js they run (standard vs enhanced) — e.g. es5-shim for IE8

Some obvious culprits that I can think of:

* a lot of standard code is written with compatibility with IE8 in mind: polyfills and helpers that are not needed for most browsers. Worth noting that [only IE11 appears in the list of top browsers](https://frontend.gutools.co.uk/analytics/quality) that hit theguardian.com... We should definitely consider updating our "cut the mustard" policy.
* many lodash helpers have an enormous amount of dependencies. Even in an optimized build, that's a lot of code. Now, lodash is an open source library meant for rookie js devs, so it has lots of bells and whistles that _we don't need_. We can do better.
* some standard modules are really huge because of careless coding. e.g. the `survey-simple` AB test required the whole `svgs` module; that's 44% of boot.js. others included modules that were not even used.

![picture 1](https://cloud.githubusercontent.com/assets/629976/15426303/8f9e7f94-1e85-11e6-9c53-b31afa5001cb.jpg)

I have gone through all the modules loaded in standard and removed as much as I could (mostly `bonzo`, and `lodash` collections functions). That's 150k worth of work removed:

![picture 2](https://cloud.githubusercontent.com/assets/629976/15426388/3a1b362e-1e86-11e6-854e-3335cd67aadb.jpg)

This is obviously a lot in a single PR and I can of course split this into more manageable chunks. But this is a start; I also plan to spend time on

* doing the same in the commercial js
* precompiling lodash templates (still a big, unnecessary part of boot.js)
* optimising the commercial code

I also think we really should update our browser compatibility strategy, have lines in the sand for standard modules, and decide on acceptable coding patterns. @ScottPainterGNM is already thinking about this; ultimately that is a discussion we should all be part of.