@uplne 2016-05-20
So basically you are 100% confident with native forEach etc right