@jbreckmckye 2016-05-20
@uplne We don't run any of the commercial scripts in non-enhanced browsers (e.g. IE8) any longer, so forEach and friends should work just fine. If not, we'll spot any errors in Sentry pretty quickly.