@gtrufitt 2016-05-20
Cool! I am certainly guilty of using the libraries instead of the native methods where I have no reason to and we can definitely do better.

If we remove these helper modules/libraries they obv need to be removed everywhere to have the effect we want but the problem I have with the pure-JS approach is having to do things like this:

```
 var body = document.getElementsByTagName('body')[0];
                         if ('classList' in body) {
                             body.classList.remove('has-membership-access-requirement');
                         } else {
                             body.className = body.className.replace('has-membership-access-requirement', '');
                        }
```

It just feels kinda error prone and though the majority of us are perfectly capable of remembering to do this, this codebase has lots of different developers working on it.

I think we need to think of *how* we tackle this - can we add some JSHint rules against the use of things like `$`, `removeClass` etc and add helpful warning/error messages that suggest the approach? Prevent the use of lodash in standard scripts?

That said the `classList` example here is moot if we move IE9 out of modern browsers and there's not a whole load of other areas where that complexity rears its head - maybe the [assign](https://github.com/guardian/frontend/pull/13005/files#diff-7b9f143f8e033237f74cd5988d18713dR24) method. Maybe we could move these into our own utils instead.

Anyway, this is great work, looking forward to seeing how much we can do going forward!

@GHaberis - You've been thinking about the same thing in the apps too, right? You've removed your micro-libraries?