@gtrufitt 2016-05-20
Actually, the `classLIst` example isn't moot is it because it's in main.js