@regiskuckaertz 2016-05-20
@gtrufitt definitely agree with all you said. That is why I think there should be specific lines in the sand for standard vs enhanced. In standard, there should be an emphasis on being as close as possible to browser engines in order to reduce overhead and increase perfs. I believe we can avoid the kind of coding you outlined 99% of the time, but yeah this excerpt exemplifies the sacrifice we'd have to do _sometimes_. If IE9 is out of the equation, we can add a `classList` polyfill and using the standard API ... (this is the only reason we're still using `bonzo` anyway).

Pragmatically speaking, a line should be drawn somewhere as to when it is ok to use a microlib in standard. I believe this is where the emphasis on recent browsers can help us move the needle in the right direction.