@GHaberis 2016-05-20
@gtrufitt I've wanted to remove micro-libraries within the app templates, but haven't gone all the way yet. I do remove them as and when I touch code where I deem them unnecessary.

I've wanted to remove them for a few reasons. Firstly for performance gains. In previous roles I've seen load times improve in hybrid apps by keeping templates as lightweight as possible by removing libraries and their dependencies. Secondly libraries are used everywhere in situations were vanilla native JS could've done the same without any additional overhead. I also happen to have come from an environment where native JS was strongly encouraged above libraries, so I'm happy using native js. However I've held back removing libraries for a few reasons...

- I don't intend to be the sole contributor to the codebase, therefore I want others input into whether removing qwery, bean, bonzo etc is something they'd be happy with if they have to start contributing. Especially as every other project I've seen here uses these libraries.
- I'd like to actually prove performance gains can be made by removing these libraries, however because they're used by virtually every module I'd need to make the changes across the entire codebase to see these gains.

On the flipside I can see the benefits of using a utility library like lodash and being able to code in a more declarative/functional way. Basically I'm at a point where I know there are certain libraries I definitely want to remove but I'm unsure about imposing a 'no libaries at all' policy!