@regiskuckaertz 2016-05-23
I also think micro libs are very useful (I'm writing [my own](https://github.com/regiskuckaertz/seier)), in that they help write concise and maintainable code that is easier to read for everyone. I just don't think lodash is the right fit for us. Given its appeal as a plug-in library that _just works_ everywhere, it cannot make many assumptions about the data it will be fed or the environment it will run in. That's the trade-off many open source projects must make if they want to spread… As @gtrufitt suggests, we could have our own simile of lodash, optimised for our unique context.

Essentially, standard and enhanced are two diametrically opposed environments: in enhanced, we have much more leeway. Degrading performance (a little) for the sake of readability/maintainability is acceptable–in standard, it is much less so. We could make the parallel with writing a device driver and an enterprise app, the former in C and the latter in ~~Java~~ Scala. I'm counting the days until we can rewrite standard in [wasm](https://webassembly.github.io/) 👯.