@gtrufitt 2016-05-23
Aside from the comments raised around polyfills, which you may or may not address in this PR, this looks good 👍