@regiskuckaertz 2016-05-23
@gtrufitt thanks 😄  one question I have is, should standard run in IE8? if yes then I just need to include es5-shim somewhere.