@sndrs 2016-05-23
sorry i've not given this the time, i will do now, but yes – current idea is that standard js will try to run in all envs, like the inline JS. obviously some things are not supported, but it will always try (i.e. not be guarded by an `isModern`-like flag.

the only difference between it and the inline JS is that inline directly affects rendering (either blocking, which we want to run before render starts, or non-blocking, which we want to runs asap on each page, but not block).