@sndrs 2016-05-23
this lgtm, but my god it's a bruiser :) 

hugely applaud the work that's gone into this and the rationale, just a couple of qs:

1. be interested to know how much faster this runs, esp on an older device? _is_ is faster?
2. this definitely needs a response to the issue of standard running everywhere, es5-shim sounds perfectly reasonable to me.
3. this had a _lot_ of changes in it, which makes it hard to review (as you point out). smaller chunks would be useful?

in terms of standard/support etc, totally no problem with some code not running where it can't e.g. localstorage is not supported, but we need to make sure anything outside a baseline check fails gracefully, and also works if it can. i.e. you could use `for` loops in place of `forEach`, `map` etc.

an es5 baseline means ~2% of current users can't run it – it seems reasonable to me that we polyfil in those circumstances, but it's worth noting too that curl will try and load the app on IE6+ – we need to know what we'll do in those circumstances.

👎 till polyfils, but 👏 the effort!