@regiskuckaertz 2016-05-23
Thanks Alex!

1. I'll grab comparison metrics on both prod and dev modes asap
2. Check out [this PR](https://github.com/guardian/frontend/pull/13031)
3. Yep I think it's better if I do it that way