@regiskuckaertz 2016-05-25
@sndrs In response to 1. I ran tests on the network front and an article both in DEV and PROD and consistently got between 400ms–1000ms improvements in how fast standard kicked in.

I'll close this PR and will release fixes in smaller batches. I think it will need a bit of communication too (e.g. stop using `svgs` in standard).